# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API PU-SDA-JATIM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parent/create": {
            "post": {
                "tags": [
                    "Parents"
                ],
                "operationId": "parentadd",
                "parameters": [
                    {
                        "name": "auhtor",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "certificate_number ",
                        "in": "query",
                        "description": "certificate_number ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "certificate_date",
                        "in": "query",
                        "description": "certificate_date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "utilization_type",
                        "in": "query",
                        "description": "utilization_type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "large",
                        "in": "query",
                        "description": "large",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "asset_value",
                        "in": "query",
                        "description": "large",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/parent/update/{id}": {
            "put": {
                "tags": [
                    "Parents"
                ],
                "operationId": "parentupdate",
                "parameters": [
                    {
                        "name": "certificate_number ",
                        "in": "query",
                        "description": "certificate_number ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "certificate_date",
                        "in": "query",
                        "description": "certificate_date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "utilization_type",
                        "in": "query",
                        "description": "utilization_type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "large",
                        "in": "query",
                        "description": "large",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "asset_value",
                        "in": "query",
                        "description": "large",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}